+++ packages/lumx-react/src/components/thumbnail/Thumbnail.stories.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 
 export default { title: 'LumX components/thumbnail/Thumbnail' };
 
-const numberKnobOtions = {
+const numberKnobOptions = {
     max: 1,
     min: -1,
     range: true,
@@ -24,10 +24,14 @@ export const defaultThumbnail = ({ theme }: { theme: Theme }) => {
             crossOrigin={select('CORS', CrossOrigin, CrossOrigin.anonymous, 'Options')}
             fillHeight={boolean('Fill Height', false, 'Options')}
             focusPoint={{
-                x: number('focusX', 0, numberKnobOtions, 'Options'),
-                y: number('focusY', 0, numberKnobOtions, 'Options'),
+                x: number('focusX', 0, numberKnobOptions, 'Options'),
+                y: number('focusY', 0, numberKnobOptions, 'Options'),
             }}
-            image={text('Url image', 'https://i.picsum.photos/id/1001/2400/1400.jpg', 'Options')}
+            image={text(
+                'Url image',
+                'https://i.picsum.photos/id/237/536/354.jpg?hmac=i0yVXW1ORpyCZpQ-CknuyV-jbtU7_x9EBQVhvT5aRr0',
+                'Options',
+            )}
             size={select(
                 'Size',
                 {
@@ -46,47 +50,6 @@ export const defaultThumbnail = ({ theme }: { theme: Theme }) => {
             isFollowingWindowSize={boolean('Update on window resize', true, 'Options')}
             resizeDebounceTime={number('Debounce time after resize', 20, undefined, 'Options')}
             variant={select<ThumbnailVariant>('Variant', ThumbnailVariant, ThumbnailVariant.squared, 'Options')}
-        />
-    );
-};
-
-export const thumbnailWithCustomImgProps = ({ theme }: { theme: Theme }) => {
-    const [image, setImage] = React.useState('https://not-found');
-    const onError = () => {
-        setImage('https://i.picsum.photos/id/1001/2400/1400.jpg');
-    };
-
-    return (
-        <Thumbnail
-            align={select<Alignment>('Alignment', Alignment, Alignment.left, 'Options')}
-            aspectRatio={select<AspectRatio>('Aspect ratio', AspectRatio, AspectRatio.square, 'Options')}
-            isCrossOriginEnabled={boolean('Enable CORS', true, 'Options')}
-            crossOrigin={select('CORS', CrossOrigin, CrossOrigin.anonymous, 'Options')}
-            fillHeight={boolean('Fill Height', false, 'Options')}
-            focusPoint={{
-                x: number('focusX', 0, numberKnobOtions, 'Options'),
-                y: number('focusY', 0, numberKnobOtions, 'Options'),
-            }}
-            image={image}
-            size={select(
-                'Size',
-                {
-                    XXS: Size.xxs,
-                    XS: Size.xs,
-                    S: Size.s,
-                    M: Size.m,
-                    L: Size.l,
-                    XL: Size.xl,
-                    XXL: Size.xxl,
-                },
-                Size.xxl,
-                'Options',
-            )}
-            theme={theme}
-            isFollowingWindowSize={boolean('Update on window resize', true, 'Options')}
-            resizeDebounceTime={number('Debounce time after resize', 20, undefined, 'Options')}
-            variant={select<ThumbnailVariant>('Variant', ThumbnailVariant, ThumbnailVariant.squared, 'Options')}
-            imgProps={{ onError }}
         />
     );
 };
@@ -98,7 +61,11 @@ export const brokenThumbnailWithFallbackProps = ({ theme }: { theme: Theme }) =>
         isCrossOriginEnabled={boolean('Enable CORS', true, 'Options')}
         crossOrigin={select('CORS', CrossOrigin, CrossOrigin.anonymous, 'Options')}
         fillHeight={boolean('Fill Height', false, 'Options')}
-        image={text('Url image', 'https://i.picsum.photos/id/1001/2400/1400.jp', 'Options')}
+        image={text(
+            'Url image',
+            'https://i.picsum.photos/id/237/536/354.jp?hmac=i0yVXW1ORpyCZpQ-CknuyV-jbtU7_x9EBQVhvT5aRr0',
+            'Options',
+        )}
         theme={theme}
         variant={select<ThumbnailVariant>('Variant', ThumbnailVariant, ThumbnailVariant.squared, 'Options')}
     />